services/p2/requirements.txt
aiohttp
python-socketio
protobuf
